remote recentering was not working in most cases
The signal slot system allows for signals to be connected to slots
with less arguments than the signal provides,
leftover args are discarded.
The new recentering position slot accepts one coordinate argument
whereas the old signals provide 3 components as individual args.
Dropping the last 2, the first arg was used to construct a coordinate
using its first ctor, which equalized all components with the arg.

also, VIEWPORT_UNDEFINED was used to ctor the normal vector for userMove

The ctor was intended to offer all math operations on vectors also
to scalars, a scalar arg could be used to construct a coordinate
which could then be used with vector operations.

The problematic ctor has been removed
and all scalar operations have been implemented seperately.