render calculation consistency project
vp vectors are are now based on the first cartesian octant
flipped v2 (so it is vps up-vector) and n (so it looks towards the camera)
v2 can now be used directly for rotations (done in the 1. cartesian octant)
but every use in slice calculations or on screen (8. octant) has to be negated
this allowed to fix #356 (skelvp orientation)

remove scaling vars from viewerState

texture coordinate calculation cleanup
the resulting texture of zy slicing was rotated and therefore needed special handling
zy now slices y first, but writes the texture in the same orientation as xy and xz
arb still slices with wrong scaling (this can be dealt with some other time)

rendered vertices are now in isotropic ›nanometer coordinates‹
the reverse scaling for point cloud coordinates is therefore removed
node coordinates are scaled from their dataset coordinates i.e. in renderSphere and renderCylinder
scaling of rendered objects, which was there to prevent spheres to be shown elongated, is not necessary anymore
segment length and node circRadius for frustum culling are also in iso coords

uniform ortho projection and look-at for all slice vps
this also allows to remove random translations

i wasn’t able to depth test the half transparent slice correctly
it uses small offsets now, which also fixes lines and points getting overlayed by it

dataset arrorws in skelvp show correctly in 3d now and have, like the boundary cylinders, lids
to prevent looking inside which would reveal the boundary line

color arb plane like ortho plane even if it is flipped