(Fix #400)
Lines and Points (L&P): Since in high and default render quality mode L&P replace only too-small-to-be-visible nodes, the replacing L&P are rendered with smallest visible size.
The formerly introduced zoom-dependent size is only used in low quality mode, which always uses L&P. This restriction fixes the problem that it was possible to replace small nodes with bigger points.

-  The too-small-to-be-visible condition now takes MSAA into account.