Register plugin function
Plugins can register themselves by passing a plugin instance and version to registerPlugin. The purpose of this is to ensure that only the newest ever loaded plugin will be kept.
K saves all successfully registered plugins into pluginDir, holds all registered plugins in registeredPlugins and running registered plugins in runningPlugins.
When loading a plugin with the same name as an already running plugin, a version check is done and the newer one is kept.
If a plugin with the same name is registered but not running, it is run first before running the newly loaded plugin to ensure that the version check will happen.